Simplify receivership and trainless corporation shares behaviour to make it easier to use in multiple games - Make the base game's init_share_pool method use the PRESIDENT_SALES_TO_MARKET constant - therefore games no longer need to override both the constant and the method to enable presidencies to be sellable - Move the shared behaviour of trainless corporation shares being worth 50% in 1860, 1862, and 18GB into a single module for improved maintainabillity - Add a test fixture for 1862 to ensure previous behaviour is maintained
