+++ lib/engine/game/g_18_carolinas/game.rb
@@ -326,10 +326,6 @@ module Engine
           end
         end
 
-        def init_share_pool
-          SharePool.new(self, allow_president_sale: self.class::PRESIDENT_SALES_TO_MARKET)
-        end
-
         def setup
           @saved_tiles = @tiles.dup
